refactor: Redis Service로 redis메서드 옮김
