+++ src/main/java/com/laydowncoding/tickitecking/global/service/RedisService.java
@@ -1,8 +1,11 @@
 package com.laydowncoding.tickitecking.global.service;
 
+import java.util.Collections;
+import java.util.List;
 import java.util.concurrent.TimeUnit;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.core.script.DefaultRedisScript;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -13,19 +16,10 @@ public class RedisService {
 
   private final RedisTemplate<String, String> redisTemplate;
 
-  @Transactional
-  public void setValues(String key, String value) {
-    redisTemplate.opsForValue().set(key, value);
-  }
-
   // 만료시간 설정 -> 자동 삭제
   @Transactional
   public void setValuesWithTimeout(String key, String value, long timeout) {
     redisTemplate.opsForValue().set(key, value, timeout, TimeUnit.DAYS);
-  }
-
-  public String getValues(String key) {
-    return redisTemplate.opsForValue().get(key);
   }
 
   @Transactional
@@ -35,5 +29,28 @@ public class RedisService {
 
   public boolean existsByKey(String key) {
     return redisTemplate.opsForValue().getOperations().hasKey(key);
+  }
+
+  public String addSet(String key, String value, Long expiredTime) {
+    StringBuffer stringBuffer = new StringBuffer();
+    stringBuffer.append("local keyExists = redis.call('exists', KEYS[1]) ");
+    stringBuffer.append("local isAdded ");
+    stringBuffer.append("if keyExists == 0 then ");
+    stringBuffer.append("    redis.call('sadd', KEYS[1], ARGV[1]) ");
+    stringBuffer.append("    redis.call('expire', KEYS[1], ARGV[2]) ");
+    stringBuffer.append("    isAdded = 1 ");
+    stringBuffer.append("else ");
+    stringBuffer.append("    isAdded = redis.call('sadd', KEYS[1], ARGV[1]) ");
+    stringBuffer.append("end ");
+    stringBuffer.append("return tostring(isAdded)");
+    String script = stringBuffer.toString();
+    DefaultRedisScript<String> luaScript = new DefaultRedisScript<>(script, String.class);
+    List<String> keys = Collections.singletonList(key);
+
+    return redisTemplate.execute(luaScript, keys, value, expiredTime.toString());
+  }
+
+  public void deleteValue(String key, String value) {
+    redisTemplate.opsForSet().remove(key, value);
   }
 }
